+++ workers/tests/integrationtests/test_api_report.py
@@ -31,7 +31,7 @@ class TestApiReport(unittest.TestCase):
         infile = os.path.join(
             os.path.dirname(os.path.abspath(__file__)),
             "fixtures",
-            "heidelberg_altstadt.geojson",
+            "heidelberg-altstadt-feature.geojson",
         )
         with open(infile, "r") as f:
             self.bpolys = geojson.load(f)